homework_28_final/requirements.txt
Django==5.1.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
setuptools==69.2.0
celery>=5.3
redis>=5.0
django-celery-beat>=2.5
aiogram==3.4.1